Implement Interface and Implementations for generator classes
