+++ molgenis-compute-core/src/main/java/org/molgenis/compute/generators/impl/CreateWorkflowGenerator.java
@@ -1,4 +1,4 @@
-package org.molgenis.compute.generators;
+package org.molgenis.compute.generators.impl;
 
 import org.apache.log4j.Logger;
