+++ molgenis-compute-core/src/main/java/org/molgenis/compute/generators/impl/DocTasksDiagramGenerator.ftl
