+++ molgenis-compute-core/src/main/java/org/molgenis/compute/generators/impl/GraphvizUtils.java
@@ -1,4 +1,4 @@
-package org.molgenis.compute.generators;
+package org.molgenis.compute.generators.impl;
 
 import java.io.File;
 import java.io.IOException;
