+++ molgenis-compute-core/src/main/java/org/molgenis/compute/urlreader/UrlReader.java
@@ -2,7 +2,7 @@ package org.molgenis.compute.urlreader;
 
 import java.io.File;
 
-import org.molgenis.compute.generators.BackendGenerator;
+import org.molgenis.compute.generators.impl.BackendGeneratorImpl;
 import org.molgenis.compute.model.impl.FoldParametersImpl;
 import org.molgenis.compute.parsers.impl.ProtocolParserImpl;
 
@@ -11,7 +11,7 @@ import org.molgenis.compute.parsers.impl.ProtocolParserImpl;
  * 
  * Called by:
  * 
- * {@link BackendGenerator} constructor
+ * {@link BackendGeneratorImpl} constructor
  * 
  * {@link ProtocolParserImpl} parse(File, String, ComputeProperties) method
  *
